+++ src/main/java/xyz/yourboykyle/secretroutes/events/OnChatReceive.java
@@ -28,6 +28,7 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
 import xyz.yourboykyle.secretroutes.Main;
 import xyz.yourboykyle.secretroutes.config.SRMConfig;
+import xyz.yourboykyle.secretroutes.dungeons.SecretUtils;
 import xyz.yourboykyle.secretroutes.utils.*;
 
 import java.lang.reflect.Field;
