+++ src/main/java/xyz/yourboykyle/secretroutes/events/OnEnterNewRoom.java
@@ -2,6 +2,8 @@
 package xyz.yourboykyle.secretroutes.events;
 
 import xyz.yourboykyle.secretroutes.Main;
+import xyz.yourboykyle.secretroutes.dungeons.Room;
+import xyz.yourboykyle.secretroutes.dungeons.SecretUtils;
 import xyz.yourboykyle.secretroutes.utils.*;
 
 import java.util.ArrayList;
